[i18n] fix reference line separator.
This commit fixes part of 5976facae92c commit:
  Refactor comment generating code for message catalogs.

The commit says "Refactor" but it changes original code behavior.
In original code, source locations are separated with " " but
in refactored code, source locations are separated with ", ".

msgmerge and po-mode that are part of the GNU gettext toolset treats
"," as part of source location. We can fix this problem by one of them:

  * Use " " instead of ", " as source locations separator
    in a "reference" line.
  * Put "reference" lines. Each "reference" line just has a source location.

This commit uses the latter because the latter has more higher readablility.

See also about "reference" line but it doesn't say about separator:
  http://www.gnu.org/s/hello/manual/gettext/PO-Files.html